Added Registration with Manage profile and permission

+++ src/main/java/com/anil/swiftBus/entity/Role.java
@@ -3,7 +3,17 @@ package com.anil.swiftBus.entity;
 import java.util.HashSet;
 import java.util.Set;
 
-import javax.persistence.*;
+import javax.persistence.CascadeType;
+import javax.persistence.Column;
+import javax.persistence.Entity;
+import javax.persistence.EnumType;
+import javax.persistence.Enumerated;
+import javax.persistence.FetchType;
+import javax.persistence.GeneratedValue;
+import javax.persistence.GenerationType;
+import javax.persistence.Id;
+import javax.persistence.OneToMany;
+import javax.persistence.Table;
 
 import com.anil.swiftBus.enums.UserType;
 
@@ -13,7 +23,7 @@ public class Role {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Integer id;
+    private Long id;
 
     @Enumerated(EnumType.STRING)
     @Column(name = "user_type",nullable = false, unique = true)
@@ -23,10 +33,10 @@ public class Role {
     private Set<RolePermission> rolePermissions = new HashSet<>();
     // Getters and setters
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 	public UserType getUserType() {
